backend/requirements.txt
Django>=1.11,<1.12
psycopg2>=2.7,<3.0
djangorestframework==3.9.2
django-filter>=1.0.1
djangorestframework-simplejwt
